refactored a lot of names and directories
