+++ src/App.jsx
@@ -1,10 +1,10 @@
 import { Link, MemoryRouter, Routes, Route } from 'react-router-dom'
 import { useDispatch, useSelector } from 'react-redux'
-import Frisbee from './views/Frisbee'
-import CrashSite from './views/CrashSite'
-import Level1Fight from './views/Level1Fight'
-import Level1Escape from './views/Level1Escape'
-import Level1QuizChoice from './views/Level1QuizChoice'
+import Frisbee from './views/Level1/Level1_001Frisbee'
+import CrashSite from './views/Level1/Level1_002CrashSite'
+import Level1Fight from './views/Level1/Level1Fight'
+import Level1Escape from './views/Level1/Level1Escape'
+import Level1QuizChoice from './views/Level1/Level1QuizChoice'
 import './App.css'
 import './styles/animations.css'
 import './styles/nav.css'
