+++ src/views/Level1/Level1Escape.jsx
@@ -1,12 +1,12 @@
 import React, { useEffect } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
-import { setPageAttribute, initializePageAttributes, selectPageAttributes } from '../store/slices/pageSlice';
-import EscapeScene from '../components/EscapeScene'; 
-import DialogPrompt from '../components/DialogPrompt'
+import { setPageAttribute, initializePageAttributes, selectPageAttributes } from '../../store/slices/pageSlice';
+import EscapeScene from '../../components/EscapeScene'; 
+import DialogPrompt from '../../components/DialogPrompt'
 import './Level1Escape.css';
-import Scene from '../components/Scene'
-import starryBackground from '../assets/starry-background.jpg'
-import Paragraph from '../components/Paragraph';
+import Scene from '../../components/Scene'
+import starryBackground from '../../assets/starry-background.jpg'
+import Paragraph from '../../components/Paragraph';
 
 const defaultAttributes = {
   playerHealth: 3,
