+++ src/views/Level1/Level1Fight.jsx
@@ -1,13 +1,13 @@
 import React, { useEffect } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
-import { setPageAttribute, initializePageAttributes, selectPageAttributes } from '../store/slices/pageSlice';
-import FightScene from '../components/FightScene'; 
-import DialogPrompt from '../components/DialogPrompt'
+import { setPageAttribute, initializePageAttributes, selectPageAttributes } from '../../store/slices/pageSlice';
+import FightScene from '../../components/FightScene'; 
+import DialogPrompt from '../../components/DialogPrompt'
 import './Level1Fight.css';
-import Scene from '../components/Scene'
-import starryBackground from '../assets/starry-background.jpg'
-import planetForeground from '../assets/planet-foreground.png'
-import Paragraph from '../components/Paragraph';
+import Scene from '../../components/Scene'
+import starryBackground from '../../assets/starry-background.jpg'
+import planetForeground from '../../assets/planet-foreground.png'
+import Paragraph from '../../components/Paragraph';
 
 const defaultAttributes = {
   playerHealth: 10,
