+++ src/views/Level1/Level1_001Frisbee.jsx
@@ -1,18 +1,18 @@
 import { useSpring, animated } from '@react-spring/web'
 import { useEffect, useRef } from 'react'
-import Lop from '../components/characters/Lop'
-import Player from '../components/characters/Player'
-import Scene from '../components/Scene'
+import Lop from '../../components/characters/Lop'
+import Player from '../../components/characters/Player'
+import Scene from '../../components/Scene'
 import { useNavigate } from 'react-router-dom'
 import { useSelector, useDispatch } from 'react-redux'
 import { useTranslation } from 'react-i18next'
-import './Frisbee.css'
-import starryBackground from '../assets/starry-background.jpg'
-import planetForeground from '../assets/planet-foreground.png'
-import spaceship from '../assets/spaceship-256.png'
-import spaceshipSound from '../assets/spaceship-flight-crash.ogg'
-import crash from '../assets/crash.png'
-import { setPageAttribute, initializePageAttributes, selectPageAttributes } from '../store/slices/pageSlice'
+import './Level1_001Frisbee.css'
+import starryBackground from '../../assets/starry-background.jpg'
+import planetForeground from '../../assets/planet-foreground.png'
+import spaceship from '../../assets/spaceship-256.png'
+import spaceshipSound from '../../assets/spaceship-flight-crash.ogg'
+import crash from '../../assets/crash.png'
+import { setPageAttribute, initializePageAttributes, selectPageAttributes } from '../../store/slices/pageSlice'
 
 
 const defaultAttributes = {
